python-bot/requirements.txt
python-telegram-bot[ext]
fastapi
uvicorn[standard]
python-dotenv
openai
pydub
requests